Propagate the labels set in the labels capacity annotation to the machine objects (and nodes)
The new annotation to set the labels for the scale from/to zero node templates are now propagated with the ones in the legacy machineset's spec.template.metadata.labels field. The propagation happens with the same order as it does in the providers APIs and in the CAPI's autoscaler implementation.